usm: tests: classification: Move HTTP2 classification test to USM linux file
